+++ ethers-contract/Cargo.toml
@@ -39,10 +39,8 @@ hex.workspace = true
 ethers-contract-abigen = { workspace = true, optional = true }
 ethers-contract-derive = { workspace = true, optional = true }
 
-# eip712
-ethers-derive-eip712 = { workspace = true, optional = true }
-
 [dev-dependencies]
+ethers-providers = { workspace = true, features = ["ws"] }
 ethers-signers.workspace = true
 ethers-solc.workspace = true
 
@@ -51,8 +49,6 @@ tokio = { workspace = true, features = ["macros"] }
 
 [features]
 default = ["abigen"]
-
-eip712 = ["ethers-derive-eip712", "ethers-core/eip712"]
 
 abigen-offline = ["ethers-contract-abigen", "ethers-contract-derive"]
 abigen = ["abigen-offline", "ethers-contract-abigen/online"]
