+++ ethers-contract/ethers-contract-abigen/Cargo.toml
@@ -33,7 +33,7 @@ quote.workspace = true
 syn = { workspace = true, features = ["full"] }
 prettyplease = "0.1.25"
 
-Inflector = "0.11"
+Inflector.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 hex.workspace = true
