+++ ethers-contract/ethers-contract-derive/Cargo.toml
@@ -28,4 +28,6 @@ proc-macro2.workspace = true
 quote.workspace = true
 syn.workspace = true
 
+Inflector.workspace = true
 hex.workspace = true
+serde_json.workspace = true
