+++ ethers-contract/src/lib.rs
@@ -59,15 +59,12 @@ pub use ethers_contract_abigen::{
 #[cfg(any(test, feature = "abigen"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "abigen")))]
 pub use ethers_contract_derive::{
-    abigen, EthAbiCodec, EthAbiType, EthCall, EthDisplay, EthError, EthEvent,
+    abigen, Eip712, EthAbiCodec, EthAbiType, EthCall, EthDisplay, EthError, EthEvent,
 };
 
 // Hide the Lazy re-export, it's just for convenience
 #[doc(hidden)]
 pub use once_cell::sync::Lazy;
-
-#[cfg(feature = "eip712")]
-pub use ethers_derive_eip712::*;
 
 // For macro expansions only, not public API.
 // See: [#2235](https://github.com/gakonst/ethers-rs/pull/2235)
