+++ ethers-contract/tests/it/main.rs
@@ -7,6 +7,8 @@ mod derive;
 
 mod contract_call;
 
+mod eip712;
+
 #[cfg(all(not(target_arch = "wasm32"), not(feature = "celo")))]
 mod common;
