--- ethers-core/ethers-derive-eip712/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "ethers-derive-eip712"
-authors = ["Ryan Tate <ryan.michael.tate@gmail.com>"]
-description = "Derive procedural macro for EIP-712 typed data"
-
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-documentation.workspace = true
-repository.workspace = true
-homepage.workspace = true
-categories.workspace = true
-keywords.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-ethers-core = { workspace = true, features = ["eip712", "macros"] }
-
-proc-macro2.workspace = true
-quote.workspace = true
-syn.workspace = true
-
-hex.workspace = true
-serde_json.workspace = true
-
-[dev-dependencies]
-ethers-contract-derive.workspace = true
