+++ ethers-core/src/types/transaction/mod.rs
@@ -5,7 +5,6 @@ pub mod eip1559;
 pub mod eip2718;
 pub mod eip2930;
 
-#[cfg(feature = "eip712")]
 pub mod eip712;
 
 pub(crate) const BASE_NUM_TX_FIELDS: usize = 9;
