+++ ethers-signers/Cargo.toml
@@ -24,7 +24,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 all-features = true
 
 [dependencies]
-ethers-core = { workspace = true, features = ["eip712"] }
+ethers-core.workspace = true
 
 # crypto
 coins-bip32 = "0.8.3"
@@ -66,7 +66,6 @@ yubihsm = { version = "0.42.0-pre.0", features = ["secp256k1", "http", "usb"], o
 
 [dev-dependencies]
 ethers-contract-derive.workspace = true
-ethers-derive-eip712.workspace = true
 
 serde_json.workspace = true
 tempfile.workspace = true
